// generated/src/app/api-models/ipa-contract-delegations.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class IpaContractDelegations {
  seqIpaContract: number;
  delegationServiceCode: string;
  delegationFlag: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}